api/requirements.txt
fastapi==0.95.0
uvicorn==0.19.0
pymongo==4.3.3
python-dotenv==1.0.0